Escape file names passed to the shell
The plugin would previously fail if it was asked to compile/view a file whose
full name included spaces or other characters that need to be quoted for the
shell.

Vim's cfile command is a little picky about the format of the filename it
takes (it won't accept a filename-with-space in single quotes, for example)
and thus this filename must also be escaped.

There may be more special characters that need to be escaped for cfile, but
the given ones should be sufficient for normal use.